+++ backend/src/main/java/com/learningassistant/backend/modules/food/repository/FoodRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -17,22 +18,22 @@ public interface FoodRepository extends JpaRepository<Food, Long> {
     /**
      * 根据地区ID查询美食
      */
-    List<Food> findByRegionId(Integer regionId);
+    List<Food> findByRegionId(Long regionId);
 
     /**
      * 根据地区ID和类别查询美食
      */
-    List<Food> findByRegionIdAndCategory(Integer regionId, String category);
+    List<Food> findByRegionIdAndCategory(Long regionId, String category);
 
     /**
      * 根据地区ID查询美食，按评分排序
      */
     @Query("SELECT f FROM Food f WHERE f.regionId = :regionId ORDER BY f.rating DESC")
-    List<Food> findByRegionIdOrderByRating(@Param("regionId") Integer regionId);
+    List<Food> findByRegionIdOrderByRating(@Param("regionId") Long regionId);
 
     /**
      * 根据价格范围查询
      */
     @Query("SELECT f FROM Food f WHERE f.regionId = :regionId AND f.priceMax <= :maxPrice")
-    List<Food> findByRegionIdAndPriceMax(@Param("regionId") Integer regionId, @Param("maxPrice") Double maxPrice);
+    List<Food> findByRegionIdAndPriceMax(@Param("regionId") Long regionId, @Param("maxPrice") BigDecimal maxPrice);
 }
